+++ Tests/Test-Cursor.ps1
@@ -11,7 +11,7 @@ $mtest = Connect-Mdbc . test test -NewCollection
 20..1 | %{@{Value = $_}} | Add-MdbcData $mtest
 
 # get 5 skipping 5
-$cursor = Get-MdbcData $mtest -Cursor -Limit 5 -Skip 5
+$cursor = Get-MdbcData $mtest -Cursor -First 5 -Skip 5
 if ($cursor.Count() -ne 20) {throw}
 if ($cursor.Size() -ne 5) {throw}
 
@@ -27,7 +27,7 @@ if ($set[-1].Value -ne 11) {throw}
 
 # get and test ordered data
 # Set* methods returns the cursor, both handy and gotcha
-$cursor = Get-MdbcData $mtest -Cursor -Limit 5 -Skip 5
+$cursor = Get-MdbcData $mtest -Cursor -First 5 -Skip 5
 $set = $cursor.SetSortOrder('Value') | New-MdbcData
 if ($set[0].Value -ne 6) {throw}
 if ($set[-1].Value -ne 10) {throw}