+++ Tests/Test-PSObject.ps1
@@ -16,8 +16,3 @@ if ($document.GetType().FullName -ne 'Mdbc.Dictionary') { throw }
 ### Many custom objects from pipeline
 $documents = Get-Process svchost | Select-Object $property | New-MdbcData
 foreach($_ in $documents) {	if ($_.GetType().FullName -ne 'Mdbc.Dictionary') { throw } }
-
-### Convert back to custom objects with a helper and manually
-$custom1 = $documents | Convert-MdbcData
-$custom2 = $documents | %{ New-Object psobject -Property $_ }
-$custom1 | Format-Table -AutoSize | Out-String